Separate MailSender and MailSenderMock classes
Rather than having the code entangled, we now have a class which will actually send mails (MailSender) and another which will store them in Redis (MailSenderMocker)
